requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libnum==1.7.1
MarkupSafe==1.1.1
Naked==0.1.31
onetimepad==1.4
pycodestyle==2.6.0
pycryptodome==3.9.9
pyDes==2.0.1
PyYAML==5.3.1
requests==2.25.1
shellescape==3.8.1
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1